rename .objects to .utils
so that it's not confused with dvc's objects
